More log changes
Logs with an explicit base, i.e., the (list 'log base x) case doesn't go through the format-application path and this doesn't always get {} wrapped as needed for latex (e.g. when in an exponent).  Manually hardcode the {} wrapping for now. Fortunately LaTeX and KaTeX are tolerant of the extra {}s.